+++ R/hypothesis_tests.R
@@ -35,6 +35,11 @@
 #' @param var.equal a logical variable indicating whether to treat the two
 #'        variances as being equal.
 #' @param conf.level confidence level of the interval.
+#' @param parallel passed to \code{multiverse::execute_multiverse()} to indicate
+#'   whether to execute the multiverse analysis in parallel. Defaults to FALSE.
+#' @param progress passed to \code{multiverse::execute_multiverse()} to indicate
+#'   whether to include a progress bar for each step of the execution. Defaults
+#'   to FALSE.
 #' @return a multiverse table displaying the t-test results as a tibble.
 #' @name t_test_mverse
 #' @importFrom rlang .data
@@ -47,7 +52,8 @@ t_test_mverse <-
            mu = 0,
            paired = FALSE,
            var.equal = FALSE,
-           conf.level = 0.95) {
+           conf.level = 0.95,
+           parallel = FALSE, progress = FALSE) {
     stopifnot(inherits(.mverse, "mverse"))
     x <- rlang::enquo(x)
     y <- rlang::enquo(y)