+++ NEWS.md
@@ -1,3 +1,20 @@
+# mverse 0.2.1
+
+## Minor patches
+
++  Fixed `multiverse_tree()` colouring by branch.
++  Added `label_hjust` and `label_vjust` optional arguments to 
+`multiverse_tree()`.
++  Updated `add_family_branch()` and `add_formula_branch()` so that they take 
+only one branch at a time. `mverse` can only save one each of the two branch 
+types.
++  Added optional arguments `parallel` and `progress` to `execute_multiverse()`.
+The arguments are passed to `multiverse::execute_multiverse()` for parallel 
+processing and progress bar display.
++  Avoid redundant `execute_multiverse()` in `summary()` functions by extracting
+model objects from multiverse environments instead of executing the multiverses 
+each time.
+
 # mverse 0.2.0
 
 ## Major changes